Remove duplicate example files and add datagrams_js server implementation
